+++ src/api/UserMessages.js
@@ -1,0 +1,11 @@
+export class UserMessages {
+
+    constructor(messages) {
+        this.messages = messages;
+    }
+
+    addMessage(messages, sender){
+
+    }
+
+}