+++ src/server/saveReplay.js
@@ -1,5 +1,5 @@
 module.exports = function (connection, req) {
-    let sql = `INSERT INTO Replays (
+    let sql = `INSERT INTO Replay (
         jsonData,
         winner,
         replay_date) VALUES (?, ?, NOW())`;
@@ -11,7 +11,7 @@ module.exports = function (connection, req) {
         if (error) return console.error("\x1b[33m" + error.message + "\x1b[0m");
         let replayId = result.insertId;
         for (let [key, value] of new Map(JSON.parse(req.players)).entries()) {
-            connection.query(`select id From Users where username = ?`, [value.username], (error, result) => {
+            connection.query(`select id From User where username = ?`, [value.username], (error, result) => {
                 if (error) return console.error("\x1b[33m" + error.message + "\x1b[0m");
                 let sql2 = `INSERT INTO User_replay (user_id, replay_id) 
                                 VALUES (?, ?);`;