+++ src/server/updateUsersRank.js
@@ -12,7 +12,7 @@ module.exports = function (connection, req) {
         else
             rankReward = rewardOnLose(value.rank, req.rank);
         connection.query(`
-                UPDATE Users 
+                UPDATE User 
                 SET 
                 rank_points = IF(rank_points + ? >= 1 AND rank_points + ? < ?, rank_points + ?, rank_points),
                 wins = IF (username = ?, wins + 1, wins)
@@ -39,8 +39,8 @@ module.exports = function (connection, req) {
 
     function rewardOnDraw(userPoints, avgRank) {
         if (userPoints < avgRank)
-            return avgRank / userPoints;
-        return -(userPoints / avgRank);
+            return rewardOnWin(userPoints, avgRank) / 3;
+        return rewardOnLose(userPoints, avgRank) / 3;
     }
 
 };